Correccion de bug al registrar y rol usuario al registarr

+++ src/main/java/risosu/it/PokeApiClient/RestController/EntrenadorController.java
@@ -43,8 +43,8 @@ public class EntrenadorController {
     }
 
     @DeleteMapping("/{idEntrenador}")
-    public ResponseEntity Delete(@PathVariable("idEntrenador") Long idEntrenador) {
-        return ResponseEntity.ok(entrenadorService.Delete(idEntrenador));
+    public ResponseEntity Delete(@PathVariable("idEntrenador") int idEntrenador) {
+        return ResponseEntity.ok(entrenadorService.Delete(Long.valueOf(idEntrenador)));
     }
 
 }
